EDSF-65: Adding database objects into new ServicingDataContext for the institutional web portal

+++ Database Repository/EdusafeServicingDB Repository/1. Tables/dbo.InstitutionsNextPaymentAndBalanceInformation.Table.sql
@@ -13,9 +13,9 @@ IF NOT EXISTS (SELECT * FROM INFORMATION_SCHEMA.TABLES WHERE TABLE_NAME = 'Insti
 			, CreatedOn datetime not null
 			, CreatedBy varchar(25) not null
 			, InstitutionsAccountNumber bigint not null
-			, NextPaymentAmount numeric not null
+			, NextPaymentAmount float not null
 			, NextPaymentDate datetime not null
-			, CurrentBalance numeric not null
+			, CurrentBalance float not null
 			, NextPaymentStatusTypeId int not null
 			CONSTRAINT PK_InstitutionsNextPaymentAndBalanceInformation_Id PRIMARY KEY CLUSTERED (Id)
 		)
