+++ Database Repository/EdusafeServicingDB Repository/1. Tables/dbo.InstitutionsPaymentHistoryEntry.Table.sql
@@ -13,7 +13,7 @@ IF NOT EXISTS (SELECT * FROM INFORMATION_SCHEMA.TABLES WHERE TABLE_NAME = 'Insti
 			, CreatedOn datetime not null
 			, CreatedBy varchar(25) not null
 			, InstitutionsAccountNumber bigint not null
-			, PaymentAmount decimal not null
+			, PaymentAmount float not null
 			, PaymentDate datetime not null
 			, PaymentComments varchar(250) null
 			CONSTRAINT PK_InstitutionsPaymentHistoryEntry_Id PRIMARY KEY CLUSTERED (Id)
