+++ Database Repository/EdusafeServicingDB Repository/1. Tables/dbo.InsureesPremiumCalculationOptionDetails.Table.sql
@@ -14,7 +14,7 @@ IF NOT EXISTS (SELECT * FROM INFORMATION_SCHEMA.TABLES WHERE TABLE_NAME = 'Insur
 			, CreatedBy varchar(25) not null
 			, InsureesPremiumCalculationOptionDetailsSetId int not null
 			, OptionTypeId int not null 
-			, OptionPercentage decimal not null 
+			, OptionPercentage float not null 
 			CONSTRAINT PK_InsureesPremiumCalculationOptionDetails_Id PRIMARY KEY CLUSTERED (Id)
 		)	
 	END
