+++ Database Repository/EdusafeServicingDB Repository/2. Stored Procedures/dbo.SP_InsertInsureesPaymentHistoryEntry.StoredProcedure.sql
@@ -6,7 +6,7 @@ GO
 
 CREATE PROCEDURE SP_InsertInstitutionsPaymentHistoryEntry
 	@AccountNumber bigint
-	, @PaymentAmount decimal 
+	, @PaymentAmount float 
 	, @PaymentDate datetime 
 	, @PaymentComments varchar(250) null
